feat: correct configs load, update docker node test & move global peers to network interface #331
