+++ .gitignore
@@ -6,3 +6,4 @@
 .idea/
 .vscode/
 saito-node*/
+.env
