--- TODO.txt
@@ -1,5 +1,0 @@
-Casual list of suggestions on improvements:
-
- - routing hops currently contain two publickeys and one signature. we can remove "from" because it is contextually available (from the sender of the first transaction, and then the *to* field. We should also be able to reduce the signatures by switching to a different/commutive signature. This would magnificently remove the size of our transactions and shrink blocksize as well.
-
-
