+++ src/networking/mod.rs
@@ -123,6 +123,4 @@ pub mod api_message;
 pub mod filters;
 pub mod handlers;
 pub mod message_types;
-pub mod network;
-pub mod peer;
 pub mod signals;
